council_election/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
mysqlclient==2.0.3
celery==5.4
redis==3.5.3
numpy==1.21.4
pandas==1.3.4
python-dotenv==1.0.1
openpyxl==3.0.9
gunicorn==20.1.0
django-cors-headers==3.10.0
django-celery-results==2.2.0